requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
ensure
joblib
mlflow
python-box
flask
flask-Cors